requirements.txt
discord.py==2.5.2
python-dotenv==1.0.0
Flask==2.3.2
